Move TaskList into task package
