+++ src/main/java/jarvis/Jarvis.java
@@ -2,6 +2,7 @@ package jarvis;
 
 import jarvis.command.Command;
 import jarvis.exception.InvalidActionException;
+import jarvis.task.TaskList;
 
 import java.util.Scanner;
